Rename exonum => exonum_client
