+++ .gitignore
@@ -1,3 +1,7 @@
+# IDE folders
+
+.vscode
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
