+++ docs/exonum.proofs.list_proof.rst
@@ -1,23 +1,23 @@
-exonum.proofs.list\_proof package
-=================================
+exonum_client.proofs.list\_proof package
+========================================
 
 Module Contents
 ---------------
 
-.. automodule:: exonum.proofs.list_proof
+.. automodule:: exonum_client.proofs.list_proof
    :members:
 
 Submodules
 ----------
 
-exonum.proofs.list\_proof.list\_proof module
---------------------------------------------
+exonum_client.proofs.list\_proof.list\_proof module
+---------------------------------------------------
 
-.. automodule:: exonum.proofs.list_proof.list_proof
+.. automodule:: exonum_client.proofs.list_proof.list_proof
    :members:
 
-exonum.proofs.list\_proof.errors module
----------------------------------------
+exonum_client.proofs.list\_proof.errors module
+----------------------------------------------
 
-.. automodule:: exonum.proofs.list_proof.errors
+.. automodule:: exonum_client.proofs.list_proof.errors
    :members:
