+++ docs/exonum.proofs.map_proof.rst
@@ -1,29 +1,29 @@
-exonum.proofs.map\_proof package
-================================
+exonum_client.proofs.map\_proof package
+=======================================
 
 Module Contents
 ---------------
 
-.. automodule:: exonum.proofs.map_proof
+.. automodule:: exonum_client.proofs.map_proof
    :members:
 
 Submodules
 ----------
 
-exonum.proofs.map\_proof.map\_proof\_builder module
----------------------------------------------------
+exonum_client.proofs.map\_proof.map\_proof\_builder module
+----------------------------------------------------------
 
-.. automodule:: exonum.proofs.map_proof.map_proof_builder
+.. automodule:: exonum_client.proofs.map_proof.map_proof_builder
    :members:
 
-exonum.proofs.map\_proof.map\_proof module
-------------------------------------------
+exonum_client.proofs.map\_proof.map\_proof module
+-------------------------------------------------
 
-.. automodule:: exonum.proofs.map_proof.map_proof
+.. automodule:: exonum_client.proofs.map_proof.map_proof
    :members:
 
-exonum.proofs.map\_proof.errors module
---------------------------------------
+exonum_client.proofs.map\_proof.errors module
+---------------------------------------------
 
-.. automodule:: exonum.proofs.map_proof.errors
+.. automodule:: exonum_client.proofs.map_proof.errors
    :members:
