+++ examples/deploy.py
@@ -1,7 +1,7 @@
 """Example Script for Deploying the Cryptocurrency Advanced Service."""
 import time
 import json
-from exonum import ExonumClient, ModuleManager
+from exonum_client import ExonumClient, ModuleManager
 
 RUST_RUNTIME_ID = 0
