+++ examples/proofs.py
@@ -1,12 +1,11 @@
 """Example of Obtaining and Verifying Proofs."""
 
 from typing import Dict, Any
-from exonum import ExonumClient, MessageGenerator, ModuleManager
-from exonum.crypto import Hash
-from exonum.proofs import (
+from exonum_client import ExonumClient, MessageGenerator, ModuleManager
+from exonum_client.crypto import Hash
+from exonum_client.proofs import (
     ListProof,
     MapProof,
-    MapProofBuilder,
     MalformedMapProofError,
     build_encoder_function,
     ListProofVerificationError,
