+++ exonum_client/module_manager.py
