+++ exonum_client/proofs/map_proof/optional_entry.py
