+++ exonum_client/proofs/map_proof/proof_path.py
