+++ exonum_client/protobuf_loader.py
